Use biSizeImage only for compressed bitmaps
This has already been implemented once with the commit 2962b580dced73208a6e6b629e017c712160170c but was reverted with the commit 5ff5c0c6ab8fe670148b2d4b52b83f33ce086c30. This change requires biSizeImage to be checked when reading a DIB and adjusted if necessary. As a result, appended bitmap bits may be interpreted as a gap to an existing color profile.